api 콜백함수 함수화 및 구조 정리
